remove_redundant_lib
